Minor Changes to Analysis.jar to use new code
I had to reorder the Output array in Geodesic.java in order to put
closer together the ones that come the the same orientation.

Added (To the best of my abilities and knowledge of your semantics)
another Algorithm. It takes the previously written Code and gives it a
outputStream inside the Analysis.java/.jar file.